drfspam/requirements.txt
Django~=3.1.13
djangorestframework~=3.13.1
requests~=2.26.0
drf_yasg